fix: Switch from gaps to borders
All previous attempts to reconcile the erroneous resizing of the grid
despite setting constraints on its size failed. However, implementing
the separation of squares as borders instead of gaps fixes this bug.

In truth, this is technically sidestepping the issue instead of solving
it outright, but the result is, in my opinion, much better than if gaps
were continued to be used even w/o the bug.